make the transition more seamless
now tot trigger to change a page is a value that goes up and when it arrives at threshhold it still goes up for a second. tI think a beter way would be to trigger it with a true statement that will stays as is even if the value linked to it stiil goup after the threshold. It is all to fix a wierd cut of a fraction of second when i jum to the next page.